requirements.txt
numpy>=1.26,<3.0
pandas>=2.1,<3.0
pyarrow>=13.0,<17.0
scikit-learn>=1.3,<2.0
xgboost>=2.0,<3.0
torch>=2.1,<2.5
matplotlib>=3.8,<4.0
PyYAML>=6.0,<7.0
typing-extensions>=4.9,<5.0
joblib>=1.3,<2.0
tqdm>=4.66,<5.0
pytest>=7.4,<9.0
mypy>=1.8,<2.0
black>=24.4,<25.0
isort>=5.12,<6.0
